// packages/cloud-auth-client/src/model/user-state.ts
/**
 * Copyright 2020 Splunk, Inc.
 *
 * Licensed under the Apache License, Version 2.0 (the "License"): you may
 * not use this file except in compliance with the License. You may obtain
 * a copy of the License at
 *
 * http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
 * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
 * License for the specific language governing permissions and limitations
 * under the License.
 */

/**
 * User state interface.
 */
export interface UserState {
    /**
     * Tenant.
     */
    tenant: string;
    
    /**
     * Region.
     */
    region: string;

    /**
     * Email address.
     */
    email?: string;

    /**
     * Invite ID.
     */
    inviteID?: string;

    /**
     * Invite tenant.
     */
    inviteTenant?: string;

    /**
     * Accept TOS parameter.
     */
    accept_tos?: string;
}